added support for I commands in first pass

+++ structs/BinaryCommand.h
@@ -1,5 +1,7 @@
 #ifndef ASSEMBLER_HW_BINARYCOMMAND_H
 #define ASSEMBLER_HW_BINARYCOMMAND_H
+#include <stdio.h>
+
 enum LineType{R, I, J, D, ASCII, E};
 
 typedef struct BinaryCommand BinaryCommand;
@@ -9,6 +11,6 @@ struct  BinaryCommand{
     char* mask; /* indicates which bits are yet to be determined. 0 - not determined */
 } ;
 
-BinaryCommand *createBinary(unsigned int size, char* payload, char* mask);
+BinaryCommand *createBinary(size_t size, char* payload, char* mask);
 
 #endif /* ASSEMBLER_HW_BINARYCOMMAND_H */
